Separate ChildProcess transport to a submodule
